# public/docs/openapi.yaml
openapi: 3.0.0
info:
  title: 'CI4 Board API'
  description: "CodeIgniter 4 기반 헤드리스 게시판 REST API\n\n## 인증\n로그인 후 발급된 `access_token`을 `Authorization: Bearer {token}` 헤더로 전달합니다.\n\n## 기본 응답 형식\n```json\n{ \"success\": true,  \"data\": {},   \"message\": \"\" }\n{ \"success\": false, \"data\": null, \"message\": \"에러 메시지\" }\n```\n\n## API 구조\n| 경로 접두사 | 설명 | 인증 |\n|------------|------|------|\n| `/api/v1/*` | 사용자·게시판·쪽지·CMS | JWT (일부 공개) |\n| `/api/admin/v1/*` | 관리자 전용 | Admin JWT (role: superadmin/manager) |"
  version: 2.0.0
servers:
  -
    url: 'http://localhost:8080'
    description: '개발 서버'
paths:
...
  /api/admin/v1/cms/menus:
    get:
      tags:
        - AdminCMS
      summary: '메뉴 트리 조회'
      operationId: 7f21d8307e6c5876e177583c4d633f53
      responses:
        '200':
          description: '메뉴 트리'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        -
          BearerAuth: []
    post:
      tags:
        - AdminCMS
      summary: '메뉴 생성'
      operationId: c7157f1d4873d43cce09a13f69c100a0
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - label
              properties:
                label:
                  type: string
                url:
                  type: string
                target:
                  type: string
                  default: _self
                  enum: [_self, _blank]
                parent_idx:
                  type: integer
                  nullable: true
                sequence:
                  type: integer
                  default: 0
                is_used:
                  type: boolean
                  default: true
              type: object
      responses:
        '201':
          description: '생성 완료'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        -
          BearerAuth: []
  /api/admin/v1/cms/menus/reorder:
    put:
      tags:
        - AdminCMS
      summary: '메뉴 순서 일괄 변경'
      description: "드래그앤드롭 순서 일괄 저장\n\nRequest Body: [{ \"idx\": 1, \"sequence\": 0, \"parent_idx\": null }, ...]"
      operationId: 862ad4f09df45b6ecac835f84743d3a9
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                properties:
                  idx: { type: integer }
                  sequence: { type: integer }
                  parent_idx: { type: integer, nullable: true }
                type: object
      responses:
        '200':
          description: '순서 저장 완료'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        -
          BearerAuth: []
  '/api/admin/v1/cms/menus/{idx}':
    put:
      tags:
        - AdminCMS
      summary: '메뉴 수정'
      operationId: 57f64375da3242009c0e9628a86190e3
      parameters:
        -
          name: idx
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              properties:
                label:
                  type: string
                url:
                  type: string
                target:
                  type: string
                  enum: [_self, _blank]
                parent_idx:
                  type: integer
                  nullable: true
                sequence:
                  type: integer
                is_used:
                  type: boolean
              type: object
      responses:
        '200':
          description: '수정 완료'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        -
          BearerAuth: []
    delete:
      tags:
        - AdminCMS
      summary: '메뉴 삭제'
      operationId: fad73f9c6281624b921e312218ef5a07
      parameters:
        -
          name: idx
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: '삭제 완료'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        -
          BearerAuth: []
...
components:
  schemas:
    ApiResponse:
      properties:
        success:
          type: boolean
        data:
          nullable: true
        message:
          type: string
          nullable: true
      type: object
...
  responses:
...
    Unauthorized:
      description: '인증이 필요합니다.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
...
  securitySchemes:
    BearerAuth:
      type: http
      description: 'JWT Bearer token'
      bearerFormat: JWT
      scheme: bearer
...